refactor: extract namespace catalog surface policy
(cherry picked from commit 0a1f7c384fb62bad9278ef04385b319f78822edc)
