+++ service/src/main/java/ai/floedb/floecat/service/catalog/impl/NamespaceServiceImpl.java
@@ -32,9 +32,9 @@ import ai.floedb.floecat.catalog.rpc.UpdateNamespaceRequest;
 import ai.floedb.floecat.catalog.rpc.UpdateNamespaceResponse;
 import ai.floedb.floecat.common.rpc.ResourceId;
 import ai.floedb.floecat.common.rpc.ResourceKind;
-import ai.floedb.floecat.metagraph.model.NamespaceNode;
 import ai.floedb.floecat.scanner.spi.CatalogOverlay;
 import ai.floedb.floecat.scanner.spi.TopologyGraph;
+import ai.floedb.floecat.service.catalog.impl.surface.CatalogSurfaceNamespaces;
 import ai.floedb.floecat.service.common.BaseServiceImpl;
 import ai.floedb.floecat.service.common.Canonicalizer;
 import ai.floedb.floecat.service.common.IdempotencyGuard;
@@ -58,8 +58,6 @@ import io.smallrye.mutiny.Uni;
 import jakarta.inject.Inject;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Base64;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -83,90 +81,8 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
 
   private static final Set<String> NAMESPACE_MUTABLE_PATHS =
       Set.of("display_name", "description", "path", "policy_ref", "properties", "catalog_id");
-  private static final String NS_TOKEN_PREFIX = "ns:";
 
   private static final Logger LOG = Logger.getLogger(NamespaceService.class);
-
-  private List<NamespaceNode> listSystemNamespaces(ResourceId catalogId) {
-    if (catalogId == null) {
-      return List.of();
-    }
-    return overlay.listSystemNamespaces(catalogId);
-  }
-
-  private static final String PATH_DELIM = "\u001F";
-
-  private enum SystemPathMatch {
-    NONE,
-    /** Exact match to an existing SYSTEM namespace path. */
-    EXACT,
-    /** A prefix of the requested path is an existing SYSTEM namespace path (i.e. under SYSTEM). */
-    UNDER_SYSTEM
-  }
-
-  /**
-   * Returns whether {@code fullPath} collides with (EXACT) or is under (UNDER_SYSTEM) a SYSTEM
-   * namespace.
-   */
-  private SystemPathMatch systemNamespacePathMatch(
-      ResourceId catalogId, List<String> fullPath, List<NamespaceNode> sysNamespaces) {
-    if (catalogId == null || fullPath == null || fullPath.isEmpty()) {
-      return SystemPathMatch.NONE;
-    }
-
-    // Normalize the candidate path once (defensive: callers should already normalize).
-    var fullNorm = new ArrayList<String>(fullPath.size());
-    for (var seg : fullPath) {
-      fullNorm.add(normalizeName(seg));
-    }
-
-    // Build a set of SYSTEM namespace full paths (normalized), encoded as a single string.
-    var sysFullPaths = new java.util.HashSet<String>();
-    for (var n : sysNamespaces) {
-      if (n == null) {
-        continue;
-      }
-
-      var leaf = n.displayName();
-      if (leaf == null || leaf.isBlank()) {
-        continue;
-      }
-
-      // Encode the SYSTEM full path with a StringBuilder to reduce allocations.
-      var sb = new StringBuilder();
-      boolean first = true;
-      for (var seg : n.pathSegments()) {
-        var s = normalizeName(seg);
-        if (!first) {
-          sb.append(PATH_DELIM);
-        }
-        sb.append(s);
-        first = false;
-      }
-      var leafNorm = normalizeName(leaf);
-      if (!first) {
-        sb.append(PATH_DELIM);
-      }
-      sb.append(leafNorm);
-
-      sysFullPaths.add(sb.toString());
-    }
-
-    // Check prefixes efficiently without allocating a new joined string each time.
-    var sb = new StringBuilder();
-    for (int i = 0; i < fullNorm.size(); i++) {
-      if (i > 0) {
-        sb.append(PATH_DELIM);
-      }
-      sb.append(fullNorm.get(i));
-
-      if (sysFullPaths.contains(sb.toString())) {
-        return (i == fullNorm.size() - 1) ? SystemPathMatch.EXACT : SystemPathMatch.UNDER_SYSTEM;
-      }
-    }
-
-    return SystemPathMatch.NONE;
-  }
 
   // ---------- RPCs ----------
 
@@ -179,172 +95,8 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                 () -> {
                   var princ = principal.get();
                   authz.require(princ, "namespace.read");
-                  final String accountId = princ.getAccountId();
-
-                  final ResourceId catalogId;
-                  final List<String> parentPath;
-
-                  if (request.hasNamespaceId()) {
-                    // Overlay is source of truth for both user and system namespace parents.
-                    var parentNode =
-                        CatalogOverlayGuards.requireVisibleNamespaceNode(
-                            overlay, request.getNamespaceId(), correlationId());
-                    catalogId = parentNode.catalogId();
-                    parentPath = append(parentNode.pathSegments(), parentNode.displayName());
-                  } else if (request.hasCatalogId()) {
-                    // Overlay is source of truth for both user and synthetic system catalogs.
-                    catalogId =
-                        CatalogOverlayGuards.requireVisibleCatalogNode(
-                                overlay, request.getCatalogId(), correlationId())
-                            .id();
-                    parentPath = new ArrayList<>(request.getPathList());
-                  } else {
-                    throw GrpcErrors.invalidArgument(
-                        correlationId(),
-                        GeneratedErrorMessages.MessageKey.SELECTOR_REQUIRED,
-                        Map.of());
-                  }
-
-                  final boolean recursive = request.getRecursive();
-                  if (request.getChildrenOnly() && recursive) {
-                    throw GrpcErrors.invalidArgument(
-                        correlationId(),
-                        null,
-                        Map.of("children_only", "true", "recursive", "true"));
-                  }
-
-                  final String namePrefix = request.getNamePrefix().trim();
-                  final List<NamespaceNode> sysNamespaces = listSystemNamespaces(catalogId);
-
-                  var pageIn = MutationOps.pageIn(request.hasPage() ? request.getPage() : null);
-                  final int want = Math.max(1, pageIn.limit);
-                  final int batch = Math.max(want * 4, 64);
-
-                  final boolean isServiceToken =
-                      pageIn.token != null && pageIn.token.startsWith(NS_TOKEN_PREFIX);
-                  final String resumeAfterRel = isServiceToken ? decodeNsToken(pageIn.token) : "";
-                  String cursor = isServiceToken ? "" : pageIn.token;
-
-                  var out = new ArrayList<Namespace>(want);
-                  String lastEmittedRel = "";
-
-                  // Phase 1: existing repo-backed pagination (unchanged)
-                  while (out.size() < want) {
-                    var next = new StringBuilder();
-                    final List<Namespace> scanned;
-                    try {
-                      scanned =
-                          namespaceRepo.list(
-                              accountId, catalogId.getId(), parentPath, batch, cursor, next);
-                    } catch (IllegalArgumentException badToken) {
-                      throw GrpcErrors.invalidArgument(
-                          correlationId(),
-                          GeneratedErrorMessages.MessageKey.PAGE_TOKEN_INVALID,
-                          Map.of("page_token", cursor));
-                    }
-
-                    if (recursive) {
-                      for (var ns : scanned) {
-                        if (!isDescendantOf(ns, parentPath)) {
-                          continue;
-                        }
-
-                        var rel = relativeQualifiedName(ns, parentPath);
-                        if (!namePrefix.isBlank() && !rel.startsWith(namePrefix)) {
-                          continue;
-                        }
-
-                        if (!resumeAfterRel.isBlank() && rel.compareTo(resumeAfterRel) <= 0) {
-                          continue;
-                        }
-
-                        out.add(ns);
-                        lastEmittedRel = rel;
-                        if (out.size() >= want) {
-                          break;
-                        }
-                      }
-                    } else {
-                      for (var ns : scanned) {
-                        if (!isImmediateChildOf(ns, parentPath)) {
-                          continue;
-                        }
-
-                        var rel = relativeQualifiedName(ns, parentPath);
-                        if (!namePrefix.isBlank() && !rel.startsWith(namePrefix)) {
-                          continue;
-                        }
-
-                        if (!resumeAfterRel.isBlank() && rel.compareTo(resumeAfterRel) <= 0) {
-                          continue;
-                        }
-
-                        out.add(ns);
-                        lastEmittedRel = rel;
-                        if (out.size() >= want) {
-                          break;
-                        }
-                      }
-                    }
-
-                    cursor = next.toString();
-                    if (cursor.isBlank()) {
-                      break;
-                    }
-
-                    if (out.size() >= want) {
-                      break;
-                    }
-                  }
-
-                  // Phase 2: only once repo cursor is exhausted, append SYSTEM namespaces.
-                  if (cursor.isBlank() && out.size() < want) {
-
-                    record SysItem(NamespaceNode node, String rel) {}
-
-                    var sysItems =
-                        sysNamespaces.stream()
-                            .filter(
-                                ns ->
-                                    recursive
-                                        ? isDescendantOf(ns, parentPath)
-                                        : isImmediateChildOf(ns, parentPath))
-                            .map(ns -> new SysItem(ns, relativeQualifiedName(ns, parentPath)))
-                            .filter(it -> namePrefix.isBlank() || it.rel().startsWith(namePrefix))
-                            .sorted(Comparator.comparing(SysItem::rel))
-                            .toList();
-
-                    for (var it : sysItems) {
-                      if (!resumeAfterRel.isBlank() && it.rel().compareTo(resumeAfterRel) <= 0) {
-                        continue;
-                      }
-                      if (out.size() >= want) {
-                        break;
-                      }
-                      out.add(toProto(it.node()));
-                      lastEmittedRel = it.rel();
-                    }
-                  }
-
-                  String nextToken = cursor;
-                  if (nextToken.isBlank() && out.size() == want) {
-                    nextToken = encodeNsToken(lastEmittedRel);
-                  }
-
-                  final int total =
-                      countNamespaces(
-                          accountId,
-                          catalogId.getId(),
-                          parentPath,
-                          namePrefix,
-                          recursive,
-                          sysNamespaces);
-
-                  var page = MutationOps.pageOut(nextToken, total);
-                  return ListNamespacesResponse.newBuilder()
-                      .addAllNamespaces(out)
-                      .setPage(page)
-                      .build();
+                  return namespaceSurface()
+                      .listNamespaces(request, princ.getAccountId(), correlationId());
                 }),
             correlationId())
         .onFailure()
@@ -353,100 +105,8 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
         .invoke(L::ok);
   }
 
-  // Includes SYSTEM namespaces from overlay in totalSize.
-  private int countNamespaces(
-      String accountId,
-      String catalogId,
-      List<String> parentPath,
-      String namePrefix,
-      boolean recursive,
-      List<NamespaceNode> sysNamespaces) {
-
-    int count = 0;
-    String cursor = "";
-    while (true) {
-      var next = new StringBuilder();
-      final List<Namespace> page;
-      try {
-        page = namespaceRepo.list(accountId, catalogId, parentPath, 1000, cursor, next);
-      } catch (IllegalArgumentException bad) {
-        throw GrpcErrors.invalidArgument(
-            correlationId(),
-            GeneratedErrorMessages.MessageKey.PAGE_TOKEN_INVALID,
-            Map.of("page_token", cursor));
-      }
-
-      for (var ns : page) {
-        boolean matchesScope =
-            recursive ? isDescendantOf(ns, parentPath) : isImmediateChildOf(ns, parentPath);
-        if (!matchesScope) {
-          continue;
-        }
-
-        if (!namePrefix.isBlank()) {
-          var rel = relativeQualifiedName(ns, parentPath);
-          if (!rel.startsWith(namePrefix)) {
-            continue;
-          }
-        }
-
-        count++;
-      }
-
-      cursor = next.toString();
-      if (cursor.isBlank()) {
-        break;
-      }
-    }
-
-    int sysCount = 0;
-
-    for (var ns : sysNamespaces) {
-      boolean matchesScope =
-          recursive ? isDescendantOf(ns, parentPath) : isImmediateChildOf(ns, parentPath);
-      if (!matchesScope) {
-        continue;
-      }
-
-      if (!namePrefix.isBlank()) {
-        var rel = relativeQualifiedName(ns, parentPath);
-        if (!rel.startsWith(namePrefix)) {
-          continue;
-        }
-      }
-
-      sysCount++;
-    }
-
-    return count + sysCount;
-  }
-
-  private static boolean isDescendantOf(Namespace ns, List<String> parentPath) {
-    return isDescendantOf(ns.getParentsList(), parentPath);
-  }
-
-  private static boolean isDescendantOf(NamespaceNode ns, List<String> parentPath) {
-    return isDescendantOf(ns.pathSegments(), parentPath);
-  }
-
-  private static boolean isDescendantOf(List<String> namespaceParentPath, List<String> parentPath) {
-    if (namespaceParentPath.size() < parentPath.size()) {
-      return false;
-    }
-    for (int i = 0; i < parentPath.size(); i++) {
-      if (!namespaceParentPath.get(i).equals(parentPath.get(i))) {
-        return false;
-      }
-    }
-    return true;
-  }
-
   private static boolean isImmediateChildOf(Namespace ns, List<String> parentPath) {
     return isImmediateChildOf(ns.getParentsList(), parentPath);
-  }
-
-  private static boolean isImmediateChildOf(NamespaceNode ns, List<String> parentPath) {
-    return isImmediateChildOf(ns.pathSegments(), parentPath);
   }
 
   private static boolean isImmediateChildOf(List<String> nsParentPath, List<String> parentPath) {
@@ -469,49 +129,6 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
     return pp;
   }
 
-  private static String relativeQualifiedName(Namespace ns, List<String> parentPath) {
-    var p = ns.getParentsList();
-    int n = parentPath.size();
-    var segs = new ArrayList<String>(p.size() - n + 1);
-
-    for (int i = n; i < p.size(); i++) {
-      segs.add(p.get(i));
-    }
-
-    if (!ns.getDisplayName().isBlank()) {
-      segs.add(ns.getDisplayName());
-    }
-
-    return String.join(".", segs);
-  }
-
-  private static String relativeQualifiedName(NamespaceNode ns, List<String> parentPath) {
-    var p = ns.pathSegments();
-    int n = parentPath.size();
-    var segs = new ArrayList<String>(Math.max(0, p.size() - n) + 1);
-
-    for (int i = n; i < p.size(); i++) {
-      segs.add(p.get(i));
-    }
-
-    if (ns.displayName() != null && !ns.displayName().isBlank()) {
-      segs.add(ns.displayName());
-    }
-
-    return String.join(".", segs);
-  }
-
-  private static Namespace toProto(NamespaceNode n) {
-    return Namespace.newBuilder()
-        .setResourceId(n.id())
-        .setCatalogId(n.catalogId())
-        .setDisplayName(n.displayName())
-        .clearParents()
-        .addAllParents(n.pathSegments())
-        .putAllProperties(n.properties())
-        .build();
-  }
-
   @Override
   public Uni<GetNamespaceResponse> getNamespace(GetNamespaceRequest request) {
     var L = LogHelper.start(LOG, "GetNamespace");
@@ -521,27 +138,7 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                 () -> {
                   var princ = principal.get();
                   authz.require(princ, "namespace.read");
-                  var nsId = request.getNamespaceId();
-
-                  // Repo first (unchanged), then overlay (system).
-                  var nsOpt = namespaceRepo.getById(nsId);
-                  if (nsOpt.isPresent()) {
-                    return GetNamespaceResponse.newBuilder().setNamespace(nsOpt.get()).build();
-                  }
-
-                  var node =
-                      overlay
-                          .resolve(nsId)
-                          .filter(NamespaceNode.class::isInstance)
-                          .map(NamespaceNode.class::cast)
-                          .orElseThrow(
-                              () ->
-                                  GrpcErrors.notFound(
-                                      correlationId(),
-                                      GeneratedErrorMessages.MessageKey.NAMESPACE,
-                                      Map.of("id", nsId.getId())));
-
-                  return GetNamespaceResponse.newBuilder().setNamespace(toProto(node)).build();
+                  return namespaceSurface().getNamespace(request, correlationId());
                 }),
             correlationId())
         .onFailure()
@@ -565,15 +162,10 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                   var spec = request.getSpec();
                   PersistedSecretPropertyValidator.validateNoGeneralMetadataSecretKeys(
                       spec.getPropertiesMap(), correlationId, "spec.properties");
-                  ensureKind(
-                      spec.getCatalogId(),
-                      ResourceKind.RK_CATALOG,
-                      "spec.catalog_id",
-                      correlationId);
                   var catalog =
-                      CatalogOverlayGuards.requireVisibleCatalogNode(
-                          overlay, spec.getCatalogId(), correlationId);
-                  CatalogOverlayGuards.rejectSystemCatalogMutation(catalog.id(), correlationId);
+                      namespaceSurface()
+                          .requireWritableCatalog(
+                              spec.getCatalogId(), "spec.catalog_id", correlationId);
                   String catalogName = catalog.displayName();
 
                   var tsNow = nowTs();
@@ -602,30 +194,11 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                   }
                   final List<String> parents = List.copyOf(normalizedParents);
                   final List<String> fullPath = new ArrayList<>(parents);
-                  final List<NamespaceNode> sysNamespaces =
-                      listSystemNamespaces(spec.getCatalogId());
                   fullPath.add(display);
 
-                  var sysMatch =
-                      systemNamespacePathMatch(spec.getCatalogId(), fullPath, sysNamespaces);
-                  if (sysMatch == SystemPathMatch.EXACT) {
-                    // Exact collision with an existing SYSTEM namespace.
-                    throw GrpcErrors.alreadyExists(
-                        correlationId,
-                        GeneratedErrorMessages.MessageKey.NAMESPACE_ALREADY_EXISTS,
-                        Map.of(
-                            "display_name", display,
-                            "catalog_id", spec.getCatalogId().getId(),
-                            "path", String.join(".", fullPath)));
-                  } else if (sysMatch == SystemPathMatch.UNDER_SYSTEM) {
-                    // Reject creating at/under a SYSTEM namespace subtree.
-                    throw GrpcErrors.permissionDenied(
-                        correlationId,
-                        GeneratedErrorMessages.MessageKey.SYSTEM_OBJECT_IMMUTABLE,
-                        Map.of(
-                            "catalog_id", spec.getCatalogId().getId(),
-                            "path", String.join(".", fullPath)));
-                  }
+                  namespaceSurface()
+                      .rejectSystemNamespacePathCollision(
+                          spec.getCatalogId(), fullPath, correlationId);
 
                   final byte[] fingerprint =
                       canonicalFingerprint(spec.getCatalogId(), parents, display, spec);
@@ -804,10 +377,7 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                   authz.require(princ, "namespace.write");
 
                   var nsId = request.getNamespaceId();
-                  ensureKind(nsId, ResourceKind.RK_NAMESPACE, "namespace_id", corr);
-
-                  // SYSTEM namespaces are immutable
-                  CatalogOverlayGuards.rejectSystemNamespaceMutation(nsId, corr);
+                  namespaceSurface().requireWritableNamespace(nsId, corr);
 
                   if (!request.hasUpdateMask() || request.getUpdateMask().getPathsCount() == 0) {
                     throw GrpcErrors.invalidArgument(
@@ -917,10 +487,7 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
                   authz.require(princ, "namespace.write");
 
                   var namespaceId = request.getNamespaceId();
-                  ensureKind(namespaceId, ResourceKind.RK_NAMESPACE, "namespace_id", correlationId);
-
-                  // SYSTEM namespaces are immutable
-                  CatalogOverlayGuards.rejectSystemNamespaceMutation(namespaceId, correlationId);
+                  namespaceSurface().requireWritableNamespace(namespaceId, correlationId);
 
                   var namespace = namespaceRepo.getById(namespaceId).orElse(null);
                   var catalogId =
@@ -1045,6 +612,10 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
     return false;
   }
 
+  private CatalogSurfaceNamespaces namespaceSurface() {
+    return new CatalogSurfaceNamespaces(namespaceRepo, overlay);
+  }
+
   private static byte[] canonicalFingerprint(
       ResourceId catalogId, List<String> parents, String display, NamespaceSpec spec) {
     return new Canonicalizer()
@@ -1096,9 +667,7 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
             corr, GeneratedErrorMessages.MessageKey.CATALOG_ID_CANNOT_CLEAR, Map.of());
       }
       var cat = spec.getCatalogId();
-      ensureKind(cat, ResourceKind.RK_CATALOG, "spec.catalog_id", corr);
-      var catalog = CatalogOverlayGuards.requireVisibleCatalogNode(overlay, cat, corr);
-      CatalogOverlayGuards.rejectSystemCatalogMutation(catalog.id(), corr);
+      namespaceSurface().requireWritableCatalog(cat, "spec.catalog_id", corr);
       b.setCatalogId(cat);
     }
 
@@ -1208,26 +777,5 @@ public class NamespaceServiceImpl extends BaseServiceImpl implements NamespaceSe
       bumpParentNamespaceMarker(beforeCat.getAccountId(), beforeCat, beforeParent);
       bumpParentNamespaceMarker(afterCat.getAccountId(), afterCat, afterParent);
     }
-  }
-
-  private static String encodeNsToken(String resumeAfterRel) {
-    if (resumeAfterRel == null || resumeAfterRel.isBlank()) {
-      return "";
-    }
-
-    return NS_TOKEN_PREFIX
-        + Base64.getUrlEncoder()
-            .withoutPadding()
-            .encodeToString(resumeAfterRel.getBytes(java.nio.charset.StandardCharsets.UTF_8));
-  }
-
-  private static String decodeNsToken(String token) {
-    if (token == null || token.isBlank() || !token.startsWith(NS_TOKEN_PREFIX)) {
-      return "";
-    }
-
-    var s = token.substring(NS_TOKEN_PREFIX.length());
-    var bytes = Base64.getUrlDecoder().decode(s);
-    return new String(bytes, java.nio.charset.StandardCharsets.UTF_8);
   }
 }
